Constraints done on new Teleop page layout
All of the constraints are done for the Teleop page's new layout and it has been tested on a tablet. I also constrained the icons for the help page ranking point key, and added a bit to the endgame description text. These are probably all the changes that are needed for the scouting meeting on Saturday, so we should be good for a little bit.